// helpers/enums/colors.ts
export enum Colors {
  White = "#fff",
  Alabaster = "#fafafb",
  AmethystSmoke = "#9797B3",
  CloudBurst = "#26254E",
  VeryLightBlue = "#DDFFFB",
  BlueDiamond = "#59DDCF",
  CoralPink = "#FF7C8A",
  SunsetOrange = "#FF5668",
  VistaBlue = "#74D2A8",
  SilverSand = "#BEBED0",
  GhostWhite = "#f8f8f9",
  Black = "#000",
}